update chunking form tests
