--- test/chunking-form/samples/namespace-tracing/_expected/amd/chunk-001dbd70.js
@@ -1,9 +1,0 @@
-define(['exports', './chunk-ba02c1b7.js'], function (exports, __chunk_1) { 'use strict';
-
-  function foo() {
-    console.log('foo');
-  }
-
-  exports.foo = foo;
-
-});
