+++ test/chunking-form/samples/namespace-tracing/_expected/amd/chunk-67ec8019.js
@@ -1,0 +1,9 @@
+define(['exports', './chunk-747bf861.js'], function (exports, __chunk_1) { 'use strict';
+
+  function bar() {
+    console.log('bar');
+  }
+
+  exports.bar = bar;
+
+});
