+++ test/chunking-form/samples/namespace-tracing/_expected/amd/chunk-71f1648d.js
@@ -1,0 +1,9 @@
+define(['exports', './chunk-747bf861.js'], function (exports, __chunk_1) { 'use strict';
+
+  function foo() {
+    console.log('foo');
+  }
+
+  exports.foo = foo;
+
+});
