--- test/chunking-form/samples/namespace-tracing/_expected/amd/chunk-7f6417c9.js
@@ -1,9 +1,0 @@
-define(['exports', './chunk-ba02c1b7.js'], function (exports, __chunk_1) { 'use strict';
-
-  function bar() {
-    console.log('bar');
-  }
-
-  exports.bar = bar;
-
-});
