+++ test/chunking-form/samples/namespace-tracing/_expected/amd/main-a.js
@@ -1,4 +1,4 @@
-define(['./chunk-001dbd70.js', './chunk-ba02c1b7.js'], function (__chunk_2, __chunk_1) { 'use strict';
+define(['./chunk-71f1648d.js', './chunk-747bf861.js'], function (__chunk_2, __chunk_1) { 'use strict';
 
 	__chunk_2.foo();
 	__chunk_1.broken();
