+++ test/chunking-form/samples/namespace-tracing/_expected/amd/main-c.js
@@ -1,4 +1,4 @@
-define(['./chunk-7f6417c9.js', './chunk-ba02c1b7.js'], function (__chunk_3, __chunk_1) { 'use strict';
+define(['./chunk-67ec8019.js', './chunk-747bf861.js'], function (__chunk_3, __chunk_1) { 'use strict';
 
 	__chunk_3.bar();
 	__chunk_1.broken();
