+++ test/chunking-form/samples/namespace-tracing/_expected/cjs/chunk-3c2b236e.js
@@ -1,6 +1,6 @@
 'use strict';
 
-require('./chunk-26e52fae.js');
+require('./chunk-fb422dac.js');
 
 function foo() {
   console.log('foo');
