+++ test/chunking-form/samples/namespace-tracing/_expected/cjs/chunk-75d7bcb4.js
@@ -1,6 +1,6 @@
 'use strict';
 
-require('./chunk-26e52fae.js');
+require('./chunk-fb422dac.js');
 
 function bar() {
   console.log('bar');
