+++ test/chunking-form/samples/namespace-tracing/_expected/cjs/main-a.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var __chunk_2 = require('./chunk-22646ab0.js');
-var __chunk_1 = require('./chunk-26e52fae.js');
+var __chunk_2 = require('./chunk-3c2b236e.js');
+var __chunk_1 = require('./chunk-fb422dac.js');
 
 __chunk_2.foo();
 __chunk_1.broken();
