+++ test/chunking-form/samples/namespace-tracing/_expected/cjs/main-c.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var __chunk_3 = require('./chunk-f1d2bdc4.js');
-var __chunk_1 = require('./chunk-26e52fae.js');
+var __chunk_3 = require('./chunk-75d7bcb4.js');
+var __chunk_1 = require('./chunk-fb422dac.js');
 
 __chunk_3.bar();
 __chunk_1.broken();
