+++ test/chunking-form/samples/namespace-tracing/_expected/es/chunk-ef0a1a76.js
@@ -1,4 +1,4 @@
-import './chunk-d20557f2.js';
+import './chunk-77a3d20f.js';
 
 function bar() {
   console.log('bar');
