+++ test/chunking-form/samples/namespace-tracing/_expected/es/main-c.js
@@ -1,5 +1,5 @@
-import { a as bar } from './chunk-32e0fcbc.js';
-import { a as broken } from './chunk-d20557f2.js';
+import { a as bar } from './chunk-ef0a1a76.js';
+import { a as broken } from './chunk-77a3d20f.js';
 
 bar();
 broken();
