+++ test/chunking-form/samples/namespace-tracing/_expected/system/chunk-daa4308b.js
@@ -1,4 +1,4 @@
-System.register(['./chunk-0cb14c61.js'], function (exports, module) {
+System.register(['./chunk-4b433a97.js'], function (exports, module) {
   'use strict';
   return {
     setters: [function () {}],
