+++ test/chunking-form/samples/namespace-tracing/_expected/system/main-a.js
@@ -1,4 +1,4 @@
-System.register(['./chunk-30de4b27.js', './chunk-0cb14c61.js'], function (exports, module) {
+System.register(['./chunk-daa4308b.js', './chunk-4b433a97.js'], function (exports, module) {
 	'use strict';
 	var foo, broken;
 	return {
