+++ test/chunking-form/samples/namespace-tracing/_expected/system/main-b.js
@@ -1,4 +1,4 @@
-System.register(['./chunk-30de4b27.js', './chunk-0f1d25f2.js', './chunk-0cb14c61.js'], function (exports, module) {
+System.register(['./chunk-daa4308b.js', './chunk-adb756c4.js', './chunk-4b433a97.js'], function (exports, module) {
 	'use strict';
 	var foo, bar, broken;
 	return {
