+++ test/chunking-form/samples/namespace-tracing/_expected/system/main-c.js
@@ -1,4 +1,4 @@
-System.register(['./chunk-0f1d25f2.js', './chunk-0cb14c61.js'], function (exports, module) {
+System.register(['./chunk-adb756c4.js', './chunk-4b433a97.js'], function (exports, module) {
 	'use strict';
 	var bar, broken;
 	return {
